# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ControlWeave API
  description: GRC platform API for managing compliance frameworks, controls, evidence, and audit logs.
  version: 2.4.3
  contact:
    name: ControlWeave Support
    url: https://controlweave.com
servers:
  - url: /
    description: Default server
paths:
...
  /api/v1/organizations/{orgId}/controls:
    get:
      summary: List controls for an organization
      operationId: listOrgControls
      tags: [Controls]
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
        - name: frameworkId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Control list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgControlsResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/cmdb/assets/{assetId}/controls:
    get:
      summary: Controls mapped to an asset
      description: >
        asset_control_mappings has existed since migration 005; these are the
        first routes to reach it. The compliance status recorded here is the
        asset's own -- a control implemented organization-wide can still be
        non-compliant on one host.
      operationId: listAssetControls
      tags: [CMDB]
      parameters:
        - { name: assetId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Mappings with control reference and framework }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    post:
      summary: Map a control to an asset
      operationId: createAssetControlMapping
      tags: [CMDB]
      parameters:
        - { name: assetId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [control_id]
              properties:
                control_id: { type: string, format: uuid }
                compliance_status:
                  type: string
                  enum: [compliant, partial, non_compliant, not_applicable]
                notes: { type: string }
      responses:
        '201': { description: Mapped }
        '400': { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/cmdb/assets/{assetId}/controls/{controlId}:
    put:
      summary: Update an asset-control mapping
      operationId: updateAssetControlMapping
      tags: [CMDB]
      parameters:
        - { name: assetId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: controlId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                compliance_status:
                  type: string
                  enum: [compliant, partial, non_compliant, not_applicable]
                last_assessed: { type: string, format: date }
                next_assessment: { type: string, format: date }
                evidence_url: { type: string }
                notes: { type: string }
      responses:
        '200': { description: Updated }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    delete:
      summary: Remove an asset-control mapping
      operationId: deleteAssetControlMapping
      tags: [CMDB]
      parameters:
        - { name: assetId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: controlId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Removed }
        '404': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: array
          items:
            type: string
        correlationId:
          type: string
...
    OrgControl:
      type: object
      required: [id, control_id, title, framework_code, status]
      properties:
        id:
          type: string
          format: uuid
        control_id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        framework_code:
          type: string
        status:
          type: string
    OrgControlsResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          required: [controls]
          properties:
            controls:
              type: array
              items:
                $ref: '#/components/schemas/OrgControl'
...